Move `addAllControlPreferencesToCategory` to ControlPreference.kt
The method should be a responsibility of the screen's class, not of the individual preference's class
